エディタリファクタリング
